Add empty member info activity
