--- app/src/main/res/layout/activity_members_list.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".activity.MembersListActivity">
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_gravity="right">
-
-        <EditText
-            android:id="@+id/et_find_member"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:hint="find member" />
-
-        <Button
-            android:id="@+id/btn_find_member"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="find"/>
-
-    </LinearLayout>
-
-    <ListView
-        android:id="@+id/member_list"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-
-</LinearLayout>
