+++ app/src/main/res/layout/activity_view_all_members.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.MemberListActivity"
+    tools:context=".activity.ViewAllMemberActivity"
     android:orientation="vertical">
 
     <LinearLayout
